LayoutTests/storage/domstorage/localstorage/quota-expected.txt

Test whether the quota is doing its job.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing window.localStorage
storage.clear()
PASS storage.length is 0
Creating 'data' which contains 64K of data
PASS data.length is 65536
Putting 'data' into 40 window.localStorage buckets.
Putting 'data' into another bucket.h
PASS Hit exception as expected
Verify that data was never inserted.
PASS storage.getItem(40) is null
Removing bucket 39.
Adding 'Hello!' into a new bucket.
PASS Insertion worked.
PASS successfullyParsed is true

TEST COMPLETE
